refactor: profileForm 위치 변경
